+++ src/main/java/madsciencemod/cards/MinatureWeapons.java
@@ -1,0 +1,57 @@
+package madsciencemod.cards;
+
+import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import madsciencemod.actions.common.ChooseAction;
+import madsciencemod.powers.MinatureWeaponsAttackPower;
+import madsciencemod.powers.MinatureWeaponsBlockPower;
+
+public class MinatureWeapons extends AbstractMadScienceCard {
+    public static final String ID = "MinatureWeapons";
+    private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+    public static final String NAME = cardStrings.NAME;
+    public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+    public static final String[] EXTENDED_DESCRIPTION = cardStrings.EXTENDED_DESCRIPTION;
+    private static final int COST = 1;
+    private static final int AMOUNT = 3;
+    private static final int UPGRADE_AMOUNT = 2;
+    private static final CardType TYPE = CardType.POWER;
+    private static final CardRarity RARITY = CardRarity.UNCOMMON;
+    private static final CardTarget TARGET = CardTarget.SELF;
+
+    public MinatureWeapons() {
+        super(ID, NAME, COST, DESCRIPTION, TYPE, RARITY, TARGET);
+        this.magicNumber = this.baseMagicNumber = AMOUNT;
+    }
+
+    @Override
+    public void use(AbstractPlayer p, AbstractMonster m) {
+        ChooseAction choice = new ChooseAction(this, m, EXTENDED_DESCRIPTION[0]);
+        choice.add(EXTENDED_DESCRIPTION[1], EXTENDED_DESCRIPTION[2], () -> {
+            AbstractDungeon.actionManager.addToBottom(new ApplyPowerAction(p, p, new MinatureWeaponsAttackPower(p, this.magicNumber), this.magicNumber));
+        });
+        choice.add(EXTENDED_DESCRIPTION[3], EXTENDED_DESCRIPTION[4], () -> {
+            AbstractDungeon.actionManager.addToBottom(new ApplyPowerAction(p, p, new MinatureWeaponsBlockPower(p, this.magicNumber), this.magicNumber));
+        });
+        AbstractDungeon.actionManager.addToBottom(choice);
+    }
+
+    @Override
+    public AbstractCard makeCopy() {
+        return new MinatureWeapons();
+    }
+
+    @Override
+    public void upgrade() {
+        if (!this.upgraded) {
+            this.upgradeName();
+            this.upgradeMagicNumber(UPGRADE_AMOUNT);
+        }
+    }
+}