+++ src/main/java/madsciencemod/powers/MinatureWeaponsBlockPower.java
@@ -1,0 +1,36 @@
+package madsciencemod.powers;
+
+import com.megacrit.cardcrawl.actions.common.GainBlockAction;
+import com.megacrit.cardcrawl.actions.utility.UseCardAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.core.AbstractCreature;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.PowerStrings;
+
+import madsciencemod.cards.AbstractTrinket;
+
+public class MinatureWeaponsBlockPower extends AbstractMadSciencePower {
+    public static final String POWER_ID = "MinatureWeaponsBlock";
+    private static final PowerStrings powerStrings = CardCrawlGame.languagePack.getPowerStrings(POWER_ID);
+    public static final String NAME = powerStrings.NAME;
+    public static final String[] DESCRIPTIONS = powerStrings.DESCRIPTIONS;
+
+    public MinatureWeaponsBlockPower(AbstractCreature owner, int amount) {
+        super(POWER_ID, NAME, owner, amount);
+        this.updateDescription();
+    }
+
+    @Override
+    public void updateDescription() {
+        this.description = DESCRIPTIONS[0] + this.amount + DESCRIPTIONS[1];
+    }
+
+    @Override
+    public void onUseCard(AbstractCard card, UseCardAction action) {
+        if (card instanceof AbstractTrinket) {
+            AbstractDungeon.actionManager.addToBottom(new GainBlockAction(this.owner, this.owner, this.amount));
+        }
+    }
+
+}